Added support for offline webpages
Right now we are trying to access a local url eg. 192.168... and the chromebox does not have internet access but it does have local network access. When we try to launch the page, it checks for network connectivity at the start which makes the app not work. Enabling offline mode would allow this function to work.